Initial 64-bit id support

+++ Telegram/NotificationService/NotificationServiceObjC/Sources/Serialization.m
@@ -16,7 +16,7 @@
     };
 }
 
-- (NSData * _Nonnull)importAuthorization:(int32_t)authId bytes:(NSData * _Nonnull)bytes {
+- (NSData * _Nonnull)importAuthorization:(int64_t)authId bytes:(NSData * _Nonnull)bytes {
     return [NSData data];
 }
