+++ submodules/TelegramCore/Sources/State/Serialization.swift
@@ -236,7 +236,7 @@ public class Serialization: NSObject, MTSerialization {
         }
     }
     
-    public func importAuthorization(_ authId: Int32, bytes: Data!) -> Data! {
+    public func importAuthorization(_ authId: Int64, bytes: Data!) -> Data! {
         return Api.functions.auth.importAuthorization(id: authId, bytes: Buffer(data: bytes)).1.makeData()
     }
